修改错误处理中间件

+++ controllers/courses.js
@@ -1,10 +1,11 @@
+const ErrorResponse = require('../utils/errorResponse')
 const courseModel = require('../modules/courses.js');
 const campsModel = require('../modules/camps.js')
 const color = require('colors');
 
 /**
  * @description 获取userid查询课程列表，有userid返回userid关联的一条课程数据，没有userid返回所有的课程数据
- * @route GET /api/v2/user/:userId/getcourses
+ * @route GET /api/v2/courses —— 获取所有课程数据 /api/v2/camps/${userId}/courses ——根据campsid查询课程
  * @access public
  */
 exports.findCourses = async(req, res, next) => {
@@ -34,33 +35,34 @@ exports.findCourses = async(req, res, next) => {
 
 /**
  * @description 创建课程
- * @route POST /api/v2/user/:userId/creatcourses
+ * @route POST /api/v2/camps/:userId/creatcourses
  * @access private
  */
 exports.CreatCourse = async(req, res, next) => {
-    let user = await campsModel.findById(req.params.userId)
+    console.log(req.body, "+++++");
+    const camp = await campsModel.findById(req.params.userId)
         // 根据userid没有查到用户信息，返回错误
-    if (!user) return next(error);
-    // 查到有用户，则向用户添加课程数据
-    const course = await courseModel.create(req.body);
+    if (!camp) return next(new ErrorResponse("camp不存在", "404"));
+    // 查到有camp，则向camp添加课程数据
+    const course = await courseModel.create(req.body)
     res.status('200').json({
         success: true,
-        data: course,
+        data: course
     });
 
 };
 
 /**
  * @description 根据id删除课程
- * @route POST /api/v2/course/:id
+ * @route DELETE /api/v2/courses/:id
  * @access private 
  */
 exports.RemoveCourses = async(req, res, next) => {
     try {
-        // if (req.params.id) {
-        //根据id删除课程
-        const course = await courseModel.findByIdAndRemove(req.params.id);
-        // }
+        if (req.params.id) {
+            //根据id删除课程
+            const course = await courseModel.findByIdAndRemove(req.params.id);
+        }
         res.status('200').json({
             success: true,
             data: {},
@@ -71,7 +73,7 @@ exports.RemoveCourses = async(req, res, next) => {
 };
 /**
  * @description 更新课程
- * @route POST /api/v2/course/:id
+ * @route POST /api/v2/courses/:id
  * @access private
  */
 exports.UpdateCourse = async(req, res, next) => {
